More work on auto-update mechanisms. Also improved TCPConnection and co a bit.

+++ src/com/stabilise/network/P255Ping.java
@@ -17,8 +17,6 @@ public class P255Ping extends Packet implements Important {
      * false if this is the return packet. */
     public boolean request;
     /** The ID of this ping transaction. */
-    // Only sent as a byte because we really don't care about ID discrepancies
-    // greater than 1; only that they're different.
     public int pingID;
     
     
@@ -32,13 +30,13 @@ public class P255Ping extends Packet implements Important {
     @Override
     public void readData(DataInStream in) throws IOException {
         request = in.readBoolean();
-        pingID = in.readByte();
+        pingID = in.readInt();
     }
     
     @Override
     public void writeData(DataOutStream out) throws IOException {
         out.writeBoolean(request);
-        out.writeByte(pingID);
+        out.writeInt(pingID);
     }
     
     @Override
